pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lang-trade"
version = "0.2.0"
description = "Local-first, resumable pipeline that turns a YouTube video into a narrated slide video."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0.0",
  "Pillow>=10.0.0",
  "python-pptx>=1.0.0",
  "yt-dlp>=2026.3.3",
  "notebooklm-mcp-cli>=0.4.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
app = "cli:main"
lang-trade = "cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["cli", "config", "exceptions", "structured_logging", "utils"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
